+++ src/components/ProjectView/LinksTab/index.jsx
@@ -11,16 +11,12 @@ const Links = React.createClass({
     return (
       <div>
         <Tabs project={project} activePath="links" />
-        <div className="project-tabs-content">
-          <div className="inner">
-            { children && project && React.cloneElement(children, {
-              project,
-              link,
-              auth,
-              authActions
-            }) }
-          </div>
-        </div>
+          { children && project && React.cloneElement(children, {
+            project,
+            link,
+            auth,
+            authActions
+          }) }
       </div>
     );
   }